producer_rabbit/requirements.txt
pika==1.3.2
Faker==24.11.0
python-dotenv==1.0.1
